requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flask-restx==1.3.0
Flask-Migrate==4.0.0
PyJWT==2.6.0
flask_testing==0.8.1
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary
flask-firebase-admin 
firebase_admin
sqlalchemy
Werkzeug==2.2.2
jwt
flasgger
flask-swagger-ui==3.36.0